requirements.txt
# Chest X-Ray Pneumonia Classifier - Requirements
# ================================================

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Experiment Tracking
tensorboard>=2.13.0
# wandb>=0.15.0  # Uncomment if using Weights & Biases

# Progress Bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# API (for inference)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Development Tools (optional)
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# pytest>=7.3.0
# pytest-asyncio>=0.21.0
